Add development mode conveniences
