+++ .env.sample
@@ -1,12 +1,12 @@
 THEME_NAME="WI Starter FSE"
 THEME_SLUG=wi-starter-fse
-THEME_DEV=true
 
 WP_VIRTUAL_HOST=${THEME_SLUG}.localhost
 WP_TABLE_PREFIX=wp_
 WP_ADMIN_USER=admin
 WP_ADMIN_PASSWORD=admin_password
 WP_LOCALE=en_US
+WP_DEVELOPMENT_MODE=theme
 
 MYSQL_ROOT_PASSWORD=root_password
 MYSQL_DATABASE=wordpress
