+++ .gitignore
@@ -166,3 +166,5 @@ dist
 /wp-content/themes/twentytwenty*
 
 /wp-content/themes/wi-starter-fse/vendor
+
+/wp-content/themes/wi-starter-fse/global-styles.css
