+++ .vscode/extensions.json
@@ -11,7 +11,9 @@
 		"esbenp.prettier-vscode",
 		"johnbillion.vscode-wordpress-hooks",
 		"devsense.phptools-vscode",
-		"editorconfig.editorconfig"
+		"editorconfig.editorconfig",
+		"stylelint.vscode-stylelint",
+		"vunguyentuan.vscode-css-variables"
 
 	],
 	// List of extensions recommended by VS Code that should not be recommended for users of this workspace.
