+++ wp-content/themes/wi-starter-fse/includes/dev/GlobalStyles.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace WI\StarterFSE\Dev;
+
+class GlobalStyles
+{
+	public function __construct()
+	{
+		add_action('wp_enqueue_scripts', array($this, 'exposeGlobalStyles'));
+	}
+
+	/**
+	 * Expose global styles when in development mode.
+	 * It will save the global styles that WP Core generates from the `theme.json`
+	 * to file that is accessible by IDEs and other tools.
+	 * This is useful for code completion and other features that require
+	 * the global styles to be available in a static file.
+	 *
+	 * @return void
+	 */
+	public function exposeGlobalStyles()
+	{
+		$stylesheet = wp_get_global_stylesheet();
+		$filename = get_template_directory() . '/global-styles.css';
+
+		file_put_contents($filename, $stylesheet);
+	}
+}
